# docs/openapi.json
{"openapi":"3.0.0","paths":{"/log/telemetry/stream":{"get":{"operationId":"SseController_telemetryStream","parameters":[{"name":"token","required":false,"in":"query","description":"JWT สำหรับ EventSource ที่ตั้ง header ไม่ได้","schema":{}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"jwt":[]}],"tags":["sse"]}},"/log/notifications/stream":{"get":{"operationId":"SseController_notificationStream","parameters":[{"name":"token","required":false,"in":"query","description":"JWT สำหรับ EventSource ที่ตั้ง header ไม่ได้","schema":{}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"jwt":[]}],"tags":["sse"]}},"/log/devices/stream":{"get":{"operationId":"SseController_deviceStream","parameters":[{"name":"token","required":false,"in":"query","description":"JWT สำหรับ EventSource ที่ตั้ง header ไม่ได้","schema":{}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"jwt":[]}],"tags":["sse"]}},"/log/stream":{"get":{"operationId":"SseController_multiStream","parameters":[{"name":"channels","required":false,"in":"query","description":"รายชื่อ channel คั่นด้วย , (telemetry | notification | device) — ไม่ระบุ = ทุก channel","schema":{"example":"device,telemetry","type":"string"}},{"name":"token","required":false,"in":"query","description":"JWT สำหรับ EventSource ที่ตั้ง header ไม่ได้","schema":{}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"jwt":[]}],"tags":["sse"]}},"/log/devices":{"post":{"operationId":"DeviceController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"serial":{"type":"string"},"ward":{"type":"string"},"staticName":{"type":"string"},"name":{"type":"string"},"status":{"type":"boolean"},"seq":{"type":"integer"},"firmware":{"type":"string"},"remark":{"type":"string"},"position":{"type":"string"},"location":{"type":"string"},"tag":{"type":"string"},"installDate":{"type":"string","format":"date-time"},"positionPic":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["devices"]},"get":{"operationId":"DeviceController_findAll","parameters":[{"name":"ward","required":false,"in":"query","schema":{"type":"array","items":{"type":"string"}}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["devices"]}},"/log/devices/{serial}":{"get":{"operationId":"DeviceController_findOne","parameters":[{"name":"serial","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["devices"]}},"/log/devices/by-name/{staticName}":{"get":{"operationId":"DeviceController_findByStaticName","parameters":[{"name":"staticName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["devices"]}},"/log/devices/by-name/{staticName}/swap":{"post":{"operationId":"DeviceController_swap","parameters":[{"name":"staticName","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwapDeviceDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"jwt":[]}],"tags":["devices"]}},"/log/devices/by-name/{staticName}/assignments":{"get":{"operationId":"DeviceController_assignmentHistory","parameters":[{"name":"staticName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["devices"]}},"/log/devices/{id}":{"put":{"operationId":"DeviceController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"ward":{"type":"string"},"staticName":{"type":"string"},"name":{"type":"string"},"status":{"type":"boolean"},"seq":{"type":"integer"},"firmware":{"type":"string"},"remark":{"type":"string"},"position":{"type":"string"},"location":{"type":"string"},"tag":{"type":"string"},"installDate":{"type":"string","format":"date-time"},"positionPic":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["devices"]}},"/log/telemetry":{"get":{"operationId":"TelemetryController_find","parameters":[{"name":"serial","required":false,"in":"query","schema":{"type":"string"}},{"name":"deviceId","required":false,"in":"query","schema":{"type":"string"}},{"name":"probeId","required":false,"in":"query","schema":{"format":"uuid","type":"string"}},{"name":"probe","required":false,"in":"query","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":1000,"default":100,"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["telemetry"]}},"/log/devices/{deviceId}/probes":{"post":{"operationId":"ProbeController_create","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProbeDto"}}}},"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["probes"]},"get":{"operationId":"ProbeController_findAllByDevice","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["probes"]}},"/log/probes/{id}":{"get":{"operationId":"ProbeController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["probes"]},"put":{"operationId":"ProbeController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProbeDto"}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["probes"]},"delete":{"operationId":"ProbeController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["probes"]}},"/log/notifications":{"post":{"operationId":"NotificationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNotificationDto"}}}},"responses":{"201":{"description":""}},"security":[{"device":[]}],"tags":["notifications"]},"get":{"operationId":"NotificationController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}},{"name":"filter","required":false,"in":"query","schema":{"type":"string"}},{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"severity","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["notifications"]}},"/log/notifications/dashboard/count":{"get":{"operationId":"NotificationController_findDashboardCount","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"jwt":[]}],"tags":["notifications"]}},"/log/notifications/unread-count":{"get":{"operationId":"NotificationController_findUnreadCount","parameters":[],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["notifications"]}},"/log/notifications/read-all":{"post":{"operationId":"NotificationController_markAllRead","parameters":[],"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["notifications"]}},"/log/notifications/{id}/read":{"patch":{"operationId":"NotificationController_markRead","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["notifications"]}},"/log/notifications/{serial}":{"get":{"operationId":"NotificationController_findBySerial","parameters":[{"name":"serial","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["notifications"]}},"/log/notifications/{id}":{"put":{"operationId":"NotificationController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNotificationDto"}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["notifications"]},"delete":{"operationId":"NotificationController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["notifications"]}},"/log/devices/by-name/{staticName}/audit":{"get":{"operationId":"AuditController_findByDevice","parameters":[{"name":"staticName","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["devices"]}},"/log/devices/{deviceId}/config":{"post":{"operationId":"DeviceConfigController_create","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDeviceConfigDto"}}}},"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["device-config"]},"get":{"operationId":"DeviceConfigController_findByDevice","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["device-config"]},"put":{"operationId":"DeviceConfigController_update","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDeviceConfigDto"}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["device-config"]},"delete":{"operationId":"DeviceConfigController_remove","parameters":[{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["device-config"]}},"/log/repairs":{"post":{"operationId":"DeviceRepairController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRepairDto"}}}},"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["device-repair"]},"get":{"operationId":"DeviceRepairController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["device-repair"]}},"/log/repairs/by-serial/{serial}":{"get":{"operationId":"DeviceRepairController_findBySerial","parameters":[{"name":"serial","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["device-repair"]}},"/log/repairs/{id}":{"get":{"operationId":"DeviceRepairController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["device-repair"]},"put":{"operationId":"DeviceRepairController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRepairDto"}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["device-repair"]},"delete":{"operationId":"DeviceRepairController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["device-repair"]}},"/log/warranties":{"post":{"operationId":"WarrantiesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWarrantyDto"}}}},"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["warranties"]},"get":{"operationId":"WarrantiesController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"expiringSoon","required":false,"in":"query","schema":{"type":"boolean"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["warranties"]}},"/log/warranties/by-serial/{serial}":{"get":{"operationId":"WarrantiesController_findBySerial","parameters":[{"name":"serial","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["warranties"]}},"/log/warranties/{id}":{"get":{"operationId":"WarrantiesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["warranties"]},"put":{"operationId":"WarrantiesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateWarrantyDto"}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["warranties"]}},"/log/firmware":{"post":{"operationId":"FirmwareController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"version":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"security":[{"jwt":[]}],"tags":["firmware"]},"get":{"operationId":"FirmwareController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["firmware"]}},"/log/firmware/latest":{"get":{"operationId":"FirmwareController_findLatest","parameters":[],"responses":{"200":{"description":""}},"tags":["firmware"]}},"/log/firmware/{id}":{"get":{"operationId":"FirmwareController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["firmware"]},"put":{"operationId":"FirmwareController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["firmware"]},"delete":{"operationId":"FirmwareController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["firmware"]}},"/log/firmware/download/{version}":{"get":{"operationId":"FirmwareController_download","parameters":[{"name":"version","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["firmware"]}},"/log/users/{actorId}/audit":{"get":{"operationId":"UserAuditController_findByActor","parameters":[{"name":"actorId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["users"]}},"/log/logday/{serialOrName}":{"get":{"operationId":"LogdayController_summary","parameters":[{"name":"serialOrName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["logday"]}},"/log/logday":{"post":{"operationId":"LogdayController_create","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"security":[{"device":[]}],"tags":["logday"]}},"/log/logday/{id}":{"put":{"operationId":"LogdayController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLogdayDto"}}}},"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["logday"]},"delete":{"operationId":"LogdayController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["logday"]}},"/log/graph/{serialOrName}":{"get":{"operationId":"GraphController_series","parameters":[{"name":"serialOrName","required":true,"in":"path","schema":{"type":"string"}},{"name":"range","required":false,"in":"query","schema":{"default":"1d","type":"string","enum":["1d","7d","30d","custom"]}},{"name":"from","required":false,"in":"query","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","schema":{"type":"string"}},{"name":"probeId","required":false,"in":"query","schema":{"format":"uuid","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"security":[{"jwt":[]}],"tags":["graph"]}},"/log/backup/months":{"get":{"operationId":"ArchiveController_listMonths","parameters":[],"responses":{"200":{"description":""}},"tags":["backup"]}},"/log/backup/exports/{month}":{"post":{"operationId":"ArchiveController_exportMonth","parameters":[{"name":"month","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["backup"]}},"/log/backup/restores/{month}":{"post":{"operationId":"ArchiveController_restoreMonth","parameters":[{"name":"month","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["backup"]},"delete":{"operationId":"ArchiveController_removeMonth","parameters":[{"name":"month","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["backup"]}},"/log/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["health"]}},"/log/mobile":{"get":{"operationId":"MobileController_findNotification","parameters":[],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["mobile"]}},"/log/mobile/{ward}":{"get":{"operationId":"MobileController_findWard","parameters":[{"name":"ward","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt":[]}],"tags":["mobile"]}}},"info":{"title":"SMtrack v2 Backend API","description":"IoT Telemetry/Log ingestion + notification backend (NestJS + Prisma + TimescaleDB + MQTT). ทุก response สำเร็จถูกห่อด้วย { success, message, data, timestamp, statusCode } ผ่าน global ResponseInterceptor และ error ถูกห่อด้วย { success: false, message, data: null } ผ่าน HttpExceptionFilter","version":"0.1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"jwt":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"User JWT (login)"},"device":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"Device auth token"}},"schemas":{"SwapDeviceDto":{"type":"object","properties":{"serial":{"type":"string","maxLength":150},"reason":{"type":"string","maxLength":255}},"required":["serial"]},"CreateProbeDto":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"channel":{"type":"string"},"tempMin":{"type":"number"},"tempMax":{"type":"number"},"humiMin":{"type":"number"},"humiMax":{"type":"number"},"tempAdj":{"type":"number"},"humiAdj":{"type":"number"},"stampTime":{"type":"string"},"doorQty":{"type":"number"},"position":{"type":"string"},"muteAlarmDuration":{"type":"string"},"doorSound":{"type":"boolean"},"doorAlarmTime":{"type":"string"},"muteDoorAlarmDuration":{"type":"string"},"notiDelay":{"type":"number"},"notiToNormal":{"type":"boolean"},"notiMobile":{"type":"boolean"},"notiRepeat":{"type":"number"},"firstDay":{"type":"string","enum":["OFF","ALL","MON","TUE","WED","THU","FRI","SAT","SUN"]},"secondDay":{"type":"string","enum":["OFF","ALL","MON","TUE","WED","THU","FRI","SAT","SUN"]},"thirdDay":{"type":"string","enum":["OFF","ALL","MON","TUE","WED","THU","FRI","SAT","SUN"]},"firstTime":{"type":"string"},"secondTime":{"type":"string"},"thirdTime":{"type":"string"}}},"UpdateProbeDto":{"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string"},"channel":{"type":"string"},"tempMin":{"type":"number"},"tempMax":{"type":"number"},"humiMin":{"type":"number"},"humiMax":{"type":"number"},"tempAdj":{"type":"number"},"humiAdj":{"type":"number"},"stampTime":{"type":"string"},"doorQty":{"type":"number"},"position":{"type":"string"},"muteAlarmDuration":{"type":"string"},"doorSound":{"type":"boolean"},"doorAlarmTime":{"type":"string"},"muteDoorAlarmDuration":{"type":"string"},"notiDelay":{"type":"number"},"notiToNormal":{"type":"boolean"},"notiMobile":{"type":"boolean"},"notiRepeat":{"type":"number"},"firstDay":{"enum":["OFF","ALL","MON","TUE","WED","THU","FRI","SAT","SUN"],"type":"string"},"secondDay":{"enum":["OFF","ALL","MON","TUE","WED","THU","FRI","SAT","SUN"],"type":"string"},"thirdDay":{"enum":["OFF","ALL","MON","TUE","WED","THU","FRI","SAT","SUN"],"type":"string"},"firstTime":{"type":"string"},"secondTime":{"type":"string"},"thirdTime":{"type":"string"}}},"CreateTelemetryDto":{"type":"object","properties":{"serial":{"type":"string"},"temp":{"type":"number"},"tempDisplay":{"type":"number"},"humidity":{"type":"number"},"humidityDisplay":{"type":"number"},"sendTime":{"type":"string"},"plug":{"type":"boolean"},"door1":{"type":"boolean"},"door2":{"type":"boolean"},"door3":{"type":"boolean"},"internet":{"type":"boolean"},"probe":{"type":"string"},"battery":{"type":"number"},"tempInternal":{"type":"number"},"extMemory":{"type":"boolean"}}},"CreateNotificationDto":{"type":"object","properties":{"serial":{"type":"string"},"message":{"type":"string"},"detail":{"type":"string"},"probe":{"type":"string"}},"required":["serial","message"]},"UpdateNotificationDto":{"type":"object","properties":{"serial":{"type":"string"},"message":{"type":"string"},"detail":{"type":"string"},"probe":{"type":"string"}}},"CreateDeviceConfigDto":{"type":"object","properties":{"dhcp":{"type":"boolean"},"ip":{"type":"string"},"mac":{"type":"string"},"subnet":{"type":"string"},"gateway":{"type":"string"},"dns":{"type":"string"},"dhcpEth":{"type":"boolean"},"ipEth":{"type":"string"},"macEth":{"type":"string"},"subnetEth":{"type":"string"},"gatewayEth":{"type":"string"},"dnsEth":{"type":"string"},"ssid":{"type":"string"},"password":{"type":"string"},"simSP":{"type":"string"},"email1":{"type":"string"},"email2":{"type":"string"},"email3":{"type":"string"},"hardReset":{"type":"string"}}},"UpdateDeviceConfigDto":{"type":"object","properties":{"dhcp":{"type":"boolean"},"ip":{"type":"string"},"mac":{"type":"string"},"subnet":{"type":"string"},"gateway":{"type":"string"},"dns":{"type":"string"},"dhcpEth":{"type":"boolean"},"ipEth":{"type":"string"},"macEth":{"type":"string"},"subnetEth":{"type":"string"},"gatewayEth":{"type":"string"},"dnsEth":{"type":"string"},"ssid":{"type":"string"},"password":{"type":"string"},"simSP":{"type":"string"},"email1":{"type":"string"},"email2":{"type":"string"},"email3":{"type":"string"},"hardReset":{"type":"string"}}},"CreateRepairDto":{"type":"object","properties":{"serial":{"type":"string"},"devName":{"type":"string"},"info":{"type":"string"},"info1":{"type":"string"},"info2":{"type":"string"},"address":{"type":"string"},"ward":{"type":"string"},"detail":{"type":"string"},"phone":{"type":"string"},"status":{"type":"string"},"warrantyStatus":{"type":"string"},"remark":{"type":"string"}},"required":["serial"]},"UpdateRepairDto":{"type":"object","properties":{"serial":{"type":"string"},"devName":{"type":"string"},"info":{"type":"string"},"info1":{"type":"string"},"info2":{"type":"string"},"address":{"type":"string"},"ward":{"type":"string"},"detail":{"type":"string"},"phone":{"type":"string"},"status":{"type":"string"},"warrantyStatus":{"type":"string"},"remark":{"type":"string"}}},"CreateWarrantyDto":{"type":"object","properties":{"serial":{"type":"string"},"devName":{"type":"string"},"product":{"type":"string"},"model":{"type":"string"},"installDate":{"type":"string"},"customerName":{"type":"string"},"customerAddress":{"type":"string"},"saleDepartment":{"type":"string"},"invoice":{"type":"string"},"expire":{"type":"string"},"status":{"type":"boolean"},"note":{"type":"string"}},"required":["serial"]},"UpdateWarrantyDto":{"type":"object","properties":{"serial":{"type":"string"},"devName":{"type":"string"},"product":{"type":"string"},"model":{"type":"string"},"installDate":{"type":"string"},"customerName":{"type":"string"},"customerAddress":{"type":"string"},"saleDepartment":{"type":"string"},"invoice":{"type":"string"},"expire":{"type":"string"},"status":{"type":"boolean"},"note":{"type":"string"}}},"UpdateLogdayDto":{"type":"object","properties":{"serial":{"type":"string"},"temp":{"type":"number"},"tempDisplay":{"type":"number"},"humidity":{"type":"number"},"humidityDisplay":{"type":"number"},"sendTime":{"type":"string"},"plug":{"type":"boolean"},"door1":{"type":"boolean"},"door2":{"type":"boolean"},"door3":{"type":"boolean"},"internet":{"type":"boolean"},"probe":{"type":"string"},"battery":{"type":"number"},"tempInternal":{"type":"number"},"extMemory":{"type":"boolean"}}}}}}